fix(desktop): added @gauzy/constants for desktop builds
